// src/enums/PollyVoice.ts
export enum PollyVoice {
  Zeina = 'Zeina',
  Zhiyu = 'Zhiyu',
  Naja = 'Naja',
  Mads = 'Mads',
  Lotte = 'Lotte',
  Ruben = 'Ruben',
  Nicole = 'Nicole',
  Russell = 'Russell',
  Amy = 'Amy',
  Emma = 'Emma',
  Brian = 'Brian',
  Aditi = 'Aditi',
  Raveena = 'Raveena',
  Ivy = 'Ivy',
  Joanna = 'Joanna',
  Kendra = 'Kendra',
  Kimberly = 'Kimberly',
  Salli = 'Salli',
  Joey = 'Joey',
  Kevin = 'Kevin',
  Geraint = 'Geraint',
  Celine = 'Celine',
  Lea = 'Lea',
  Mathieu = 'Mathieu',
  Chantal = 'Chantal',
  Marlene = 'Marlene',
  Vicki = 'Vicki',
  Hans = 'Hans',
  Dora = 'Dora',
  Karl = 'Karl',
  Carla = 'Carla',
  Bianca = 'Bianca',
  Giorgio = 'Giorgio',
  Mizuki = 'Mizuki',
  Takumi = 'Takumi',
  Seoyeon = 'Seoyeon',
  Liv = 'Liv',
  Ewa = 'Ewa',
  Maja = 'Maja',
  Jacek = 'Jacek',
  Jan = 'Jan',
  Camila = 'Camila',
  Vitoria = 'Vitoria',
  Ricardo = 'Ricardo',
  Ines = 'Ines',
  Cristiano = 'Cristiano',
  Carmen = 'Carmen',
  Tatyana = 'Tatyana',
  Maxim = 'Maxim',
  Conchita = 'Conchita',
  Lucia = 'Lucia',
  Enrique = 'Enrique',
  Mia = 'Mia',
  Lupe = 'Lupe',
  Penelope = 'Penelope',
  Miguel = 'Miguel',
  Astrid = 'Astrid',
  Filiz = 'Filiz',
  Gwyneth = 'Gwyneth',
}

export enum PollyLanguage {
  EnglishUS = 'en-US',
  SpanishUS = 'es-US',
  SpanishES = 'es-ES',
  SpanishMX = 'es-MX',
}

export enum PollyOutputFormat {
  Mp3 = 'mp3',
  Ogg = 'ogg_vorbis',
  Pcm = 'pcm',
}

export type Gender = 'Female' | 'Male' | 'Female (child)' | 'Male (child)';